pyproject.toml
[project]
name = "interview-bot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "flashrank>=0.2.10",
    "google-generativeai>=0.8.6",
    "langchain==1.0.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.24",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-qdrant>=1.1.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "presidio-analyzer>=2.2.362",
    "presidio-anonymizer>=2.2.362",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "rank-bm25>=0.2.2",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "spacy>=3.8.14",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.56.0",
    "uvicorn>=0.42.0",
]
